fixed db issues

+++ MediaLibrary/MediaLibraryDAL/Partials/PodcastItem.cs
@@ -16,7 +16,7 @@ namespace MediaLibraryDAL.DbContexts
             Title = title;
             Url = url;
             Description = description;
-            Length = length;
+            Length = (int)length;
             PublishDate = publishDate;
             PodcastId = podcastId;
         }
